# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.0.0",
    "title": "Web Tickets API",
    "description": "REST API for web tickets"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http"],
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/login": {
      "post": {
        "tags": ["Users"],
        "description": "Logs the user into the system",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "User that we want to login",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "description": "The user email"
                },
                "password": {
                  "type": "number",
                  "description": "The user password"
                }
              }
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User logged in",
            "schema": {
              "type": "object",
              "properties": {
                "auth": {
                  "type": "boolean"
                },
                "token": {
                  "type": "string",
                  "description": "The user token"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": ["Users"],
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "description": "The user first name"
                },
                "lastName": {
                  "type": "string",
                  "description": "The user last name"
                },
                "email": {
                  "type": "string",
                  "description": "The user email"
                },
                "password": {
                  "type": "number",
                  "description": "The user password"
                }
              }
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "type": "object",
              "properties": {
                "auth": {
                  "type": "boolean"
                },
                "token": {
                  "type": "string",
                  "description": "The user token"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}